fix(firmware): align LQR base layer with RL sim and fix protocol parsing bugs
LQR controller alignment (lqr_controller.h/c, main.c, pin_config.h):
- Remove position integrator, anti-windup, target_velocity, target_yaw_rate
  (RL sim holds x=0, no position loop in base layer)
- Rewrite control law to match kuafu_mjx_env._lqr_balance line-by-line:
  F = -(K·[0,θ,ẋ,θ̇]), τ_pitch = F·R/2
  conditional yaw damping: τ_diff = clip(-0.05·ωz, ±0.55) when |ωz|<0.3
  residual overlay: τ_L = τ_pitch + τ_diff + Δτ_L×0.55
- Call LQR once per 250Hz cycle (slot 0 only) instead of twice (slot 0+1)
- Add WHEEL_RADIUS_M, TAU_WHEEL_RATED, YAW_KD, YAW_DAMP_THRESH constants

ST3215 protocol fixes (st3215.c):
- sync_write_pos: clamp position to [0,4095] instead of sign-bit encoding
- read_state: position as uint16, speed/current as int16 two's complement,
  load uses bit10 sign with safe bitmask
- Clean up verbose comments, remove dead RADS_TO_TICK macro, fix tx_buf size

Mahony simplification (mahony.c):
- Remove custom asinff helper, inline clamp+asinf
- Pitch formula matches RL: arcsin(clip(2*(qw*qy-qx*qz)))

Unify byte-assembly pattern (bmi088.c, ddsm315.c, pi_link.c):
- Use ((uint16_t)hi<<8)|lo everywhere to eliminate integer promotion issues

Other:
- safety_state.c: add (void)gyro_y_rads to suppress unused warning
- Verify bmi088 temperature read chain is complete

+++ stm32_firmware/Core/Src/bmi088.c
@@ -88,10 +88,10 @@ int bmi088_read_accel(BMI088_t *imu) {
         return -1;
     }
 
-    /* Combine raw values and scale to m/s^2 */
-    int16_t raw_x = (int16_t)((buffer[1] << 8) | buffer[0]);
-    int16_t raw_y = (int16_t)((buffer[3] << 8) | buffer[2]);
-    int16_t raw_z = (int16_t)((buffer[5] << 8) | buffer[4]);
+    /* Combine raw values (two's complement int16) and scale to m/s^2 */
+    int16_t raw_x = (int16_t)(((uint16_t)buffer[1] << 8) | buffer[0]);
+    int16_t raw_y = (int16_t)(((uint16_t)buffer[3] << 8) | buffer[2]);
+    int16_t raw_z = (int16_t)(((uint16_t)buffer[5] << 8) | buffer[4]);
 
     imu->accel[0] = (float)raw_x * ACCEL_24G_SCALE;
     imu->accel[1] = (float)raw_y * ACCEL_24G_SCALE;
@@ -108,10 +108,10 @@ int bmi088_read_gyro(BMI088_t *imu) {
         return -1;
     }
 
-    /* Combine raw values and scale to rad/s */
-    int16_t raw_x = (int16_t)((buffer[1] << 8) | buffer[0]);
-    int16_t raw_y = (int16_t)((buffer[3] << 8) | buffer[2]);
-    int16_t raw_z = (int16_t)((buffer[5] << 8) | buffer[4]);
+    /* Combine raw values (two's complement int16) and scale to rad/s */
+    int16_t raw_x = (int16_t)(((uint16_t)buffer[1] << 8) | buffer[0]);
+    int16_t raw_y = (int16_t)(((uint16_t)buffer[3] << 8) | buffer[2]);
+    int16_t raw_z = (int16_t)(((uint16_t)buffer[5] << 8) | buffer[4]);
 
     imu->gyro[0] = (float)raw_x * GYRO_2000_SCALE;
     imu->gyro[1] = (float)raw_y * GYRO_2000_SCALE;
